+++ libs/products/src/lib/containers/product-detail/product-detail.component.ts
@@ -1,4 +1,5 @@
-import { Component, OnInit, Input } from '@angular/core';
+import { Component, Output, Input, EventEmitter } from '@angular/core';
+import { faEdit } from '@fortawesome/free-solid-svg-icons';
 
 @Component({
   selector: 'demo-app-product-detail',
@@ -7,5 +8,11 @@ import { Component, OnInit, Input } from '@angular/core';
 })
 export class ProductDetailComponent {
   @Input() selectedProduct: any;
+  @Output() edit = new EventEmitter<boolean>();
+  faEdit = faEdit;
 
+  onEdit() {
+    console.log('edit1');
+    this.edit.emit(true);
+  }
 }